refactor: move zoom constants into configuration
