+++ ground/src/main/java/com/google/android/ground/ui/map/MapController.kt
@@ -15,10 +15,10 @@
  */
 package com.google.android.ground.ui.map
 
+import com.google.android.ground.Config.DEFAULT_LOI_ZOOM_LEVEL
 import com.google.android.ground.model.geometry.Point
 import com.google.android.ground.repository.SurveyRepository
 import com.google.android.ground.rx.annotations.Hot
-import com.google.android.ground.ui.home.mapcontainer.HomeScreenMapContainerViewModel.Companion.DEFAULT_LOI_ZOOM_LEVEL
 import com.google.android.ground.ui.map.gms.toPoint
 import io.reactivex.BackpressureStrategy
 import io.reactivex.Flowable
